move ColumnRegExpMetric to a separate module, fix visual, add unittests
