+++ src/evidently/metrics/__init__.py
@@ -3,7 +3,7 @@ from .classification_performance_metrics import ClassificationPerformanceMetrics
 from .classification_performance_metrics import ClassificationPerformanceMetricsThreshold
 from .classification_performance_metrics import ClassificationPerformanceMetricsTopK
 from .data_drift_metrics import DataDriftMetrics
-from .data_integrity_metrics import ColumnRegExpMetric
+from .data_integrity.column_regexp_metric import ColumnRegExpMetric
 from .data_integrity_metrics import DataIntegrityMetrics
 from .data_integrity_metrics import DataIntegrityNullValuesMetrics
 from .data_quality_metrics import DataQualityCorrelationMetrics
