+++ tests/metrics/test_data_integrity_metrics.py
@@ -2,13 +2,10 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from evidently.metrics import ColumnRegExpMetric
 from evidently.metrics import DataIntegrityMetrics
 from evidently.metrics import DataIntegrityNullValuesMetrics
 from evidently.metrics.base_metric import InputData
 from evidently.metrics.base_metric import Metric
-from evidently.metrics.data_integrity_metrics import DataIntegrityValueByRegexpMetricResult
-from evidently.metrics.data_integrity_metrics import DataIntegrityValueByRegexpStat
 from evidently.pipeline.column_mapping import ColumnMapping
 from evidently.report import Report
 
@@ -33,85 +30,9 @@ def test_data_integrity_metrics() -> None:
 
 
 @pytest.mark.parametrize(
-    "current_data, reference_data, column_name, reg_exp, expected_result",
-    (
-        (
-            pd.DataFrame(
-                {
-                    "category_feature": ["3", "a", "b5", "a", np.nan],
-                    "target": [1, 2, 1, 2, 1],
-                    "prediction": [1, 1, 1, 2, 2],
-                }
-            ),
-            None,
-            "category_feature",
-            r".*\d+.*",
-            DataIntegrityValueByRegexpMetricResult(
-                column_name="category_feature",
-                reg_exp=r".*\d+.*",
-                current=DataIntegrityValueByRegexpStat(
-                    number_of_matched=2,
-                    number_of_not_matched=2,
-                    number_of_rows=5,
-                    table_of_matched={"3": 1, "b5": 1},
-                    table_of_not_matched={"a": 2},
-                ),
-                reference=None,
-            ),
-        ),
-        (
-            pd.DataFrame(
-                {
-                    "feature": [" a", "a", "\tb", np.nan, np.nan],
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "feature": ["a", "a", "c"],
-                }
-            ),
-            "feature",
-            r"^\s+.*",
-            DataIntegrityValueByRegexpMetricResult(
-                column_name="feature",
-                reg_exp=r"^\s+.*",
-                current=DataIntegrityValueByRegexpStat(
-                    number_of_matched=2,
-                    number_of_not_matched=1,
-                    number_of_rows=5,
-                    table_of_matched={" a": 1, "\tb": 1},
-                    table_of_not_matched={"a": 1},
-                ),
-                reference=DataIntegrityValueByRegexpStat(
-                    number_of_matched=0,
-                    number_of_not_matched=3,
-                    number_of_rows=3,
-                    table_of_matched={},
-                    table_of_not_matched={"a": 2, "c": 1},
-                ),
-            ),
-        ),
-    ),
-)
-def test_data_integrity_value_by_regexp_metric(
-    current_data: pd.DataFrame,
-    reference_data: pd.DataFrame,
-    column_name: str,
-    reg_exp: str,
-    expected_result: DataIntegrityValueByRegexpMetricResult,
-) -> None:
-    metric = ColumnRegExpMetric(column_name=column_name, reg_exp=reg_exp)
-    result = metric.calculate(
-        data=InputData(current_data=current_data, reference_data=reference_data, column_mapping=ColumnMapping())
-    )
-    assert result == expected_result
-
-
-@pytest.mark.parametrize(
     "metric_object",
     (
         DataIntegrityMetrics(),
-        ColumnRegExpMetric(column_name="feature", reg_exp=r".*a.*"),
         DataIntegrityNullValuesMetrics(null_values=[None]),
     ),
 )
